Moves pipeline submodule to a subpackage
