+++ podpac/core/pipeline/__init__.py
@@ -1,0 +1,3 @@
+from podpac.core.pipeline.pipeline import Pipeline, PipelineError, PipelineNode
+from podpac.core.pipeline.output import Output, NoOutput, FileOutput, ImageOutput
+from podpac.core.pipeline.util import make_pipeline_definition
